backend/request-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy[asyncio]==2.0.31
aiosqlite==0.20.0
pydantic==2.8.0
httpx==0.27.0
grpcio==1.66.0
grpcio-tools==1.66.0
